refactor: move main page files to the separate folder
