+++ rs-class-components-app/src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import { MainPage } from './pages/Main.tsx';
+import { MainPage } from './pages/main/Main.tsx';
 import { ErrorBoundary } from './features/error-boundary/ErrorBoundary.tsx';
 import { BrowserRouter, Route, Routes } from 'react-router';
 import { Path } from './enums/Path.ts';
